Resolve circular dependency and avoid using reserved keyword 'abstract'
