+++ src/js/Properties.js
@@ -806,60 +806,23 @@ axs.properties.getAllProperties = function(node) {
     return allProperties;
 };
 
-(function() {
-    /**
-     * Helper for implicit semantic functionality.
-     * Can be made part of the public API if need be.
-     * @param {Element} element
-     * @return {?axs.constants.HtmlInfo}
-     */
-    function getHtmlInfo(element) {
-        if (!element)
-            return null;
-        var tagName = element.tagName;
-        if (!tagName)
-            return null;
-        tagName = tagName.toUpperCase();
-        var infos = axs.constants.TAG_TO_IMPLICIT_SEMANTIC_INFO[tagName];
-        if (!infos || !infos.length)
-            return null;
-        var defaultInfo = null;  // will contain the info with no specific selector if no others match
-        for (var i = 0, len = infos.length; i < len; i++) {
-            var htmlInfo = infos[i];
-            if (htmlInfo.selector) {
-                if (axs.browserUtils.matchSelector(element, htmlInfo.selector))
-                    return htmlInfo;
-            } else {
-                defaultInfo = htmlInfo;
-            }
-        }
-        return defaultInfo;
-    }
+/**
+ * @param {Element} element
+ * @return {string} role
+ */
+axs.properties.getImplicitRole = function(element) {
+    return axs.utils.getImplicitRole(element);
+}
 
-    /**
-     * @param {Element} element
-     * @return {string} role
-     */
-    axs.properties.getImplicitRole = function(element) {
-        var htmlInfo = getHtmlInfo(element);
-        if (htmlInfo)
-            return htmlInfo.role;
-        return '';
-    };
-
-    /**
-     * Determine if this element can take ANY ARIA attributes including roles, state and properties.
-     * If false then even global attributes should not be used.
-     * @param {Element} element
-     * @return {boolean}
-     */
-    axs.properties.canTakeAriaAttributes = function(element) {
-        var htmlInfo = getHtmlInfo(element);
-        if (htmlInfo)
-            return !htmlInfo.reserved;
-        return true;
-    };
-})();
+/**
+ * Determine if this element can take ANY ARIA attributes including roles, state and properties.
+ * If false then even global attributes should not be used.
+ * @param {Element} element
+ * @return {boolean}
+ */
+axs.properties.canTakeAriaAttributes = function(element) {
+    return axs.utils.canTakeAriaAttributes(element);
+}
 
 /**
  * This lists the ARIA attributes that are supported implicitly by native properties of this element.
@@ -890,38 +853,12 @@ axs.properties.getNativelySupportedAttributes = function(element) {
     return result;
 };
 
-(function() {
-    var roleToSelectorCache = {};  // performance optimization, cache results from getSelectorForRole
-
-    /**
-     * Build a selector that will match elements which implicity or explicitly have this role.
-     * Note that the selector will probably not look elegant but it will work.
-     * @param {string} role
-     * @return {string} selector
-     */
-    axs.properties.getSelectorForRole = function(role) {
-        if (!role)
-            return '';
-        if (roleToSelectorCache[role] && roleToSelectorCache.hasOwnProperty(role))
-            return roleToSelectorCache[role];
-        var selectors = ['[role="' + role + '"]'];
-        var tagNames = Object.keys(/** @type {!Object} */(axs.constants.TAG_TO_IMPLICIT_SEMANTIC_INFO));
-        tagNames.forEach(function(tagName) {
-            var htmlInfos = axs.constants.TAG_TO_IMPLICIT_SEMANTIC_INFO[tagName];
-            if (htmlInfos && htmlInfos.length) {
-                for (var i = 0; i < htmlInfos.length; i++) {
-                    var htmlInfo = htmlInfos[i];
-                    if (htmlInfo.role === role) {
-                        if (htmlInfo.selector) {
-                            selectors[selectors.length] = htmlInfo.selector;
-                        } else {
-                            selectors[selectors.length] = tagName;  // Selectors API is not case sensitive.
-                            break;  // No need to continue adding selectors since we will match the tag itself.
-                        }
-                    }
-                }
-            }
-        });
-        return (roleToSelectorCache[role] = selectors.join(','));
-    };
-})();
+/**
+ * Build a selector that will match elements which implicity or explicitly have this role.
+ * Note that the selector will probably not look elegant but it will work.
+ * @param {string} role
+ * @return {string} selector
+ */
+axs.properties.getSelectorForRole = function(role) {
+  return axs.utils.getSelectorForRole(role);
+}
