+++ test/js/utils-test.js
@@ -223,7 +223,7 @@ module("getRoles", {
 
 test("getRoles on element with valid role.", function() {
     for (var role in axs.constants.ARIA_ROLES) {
-        if (axs.constants.ARIA_ROLES.hasOwnProperty(role) && !axs.constants.ARIA_ROLES[role].abstract) {
+        if (axs.constants.ARIA_ROLES.hasOwnProperty(role) && !axs.constants.ARIA_ROLES[role]['abstract']) {
             var appliedRole = { name: role, valid: true, details: axs.constants.ARIA_ROLES[role] };
             var expected = {
                 valid: true,
@@ -277,7 +277,7 @@ test("getRoles on element with no role and options.implicit.", function() {
 
 test("getRoles on element with abstract role.", function() {
     for (var role in axs.constants.ARIA_ROLES) {
-        if (axs.constants.ARIA_ROLES.hasOwnProperty(role) && axs.constants.ARIA_ROLES[role].abstract) {
+        if (axs.constants.ARIA_ROLES.hasOwnProperty(role) && axs.constants.ARIA_ROLES[role]['abstract']) {
             var expected = {
                 valid: false,
                 roles: [{ name: role, valid: false }]
@@ -296,7 +296,7 @@ test("getRoles on element with abstract role.", function() {
      * @return The 'role detail' object.
      */
     function createExpectedRoleObject(role) {
-        var valid = (axs.constants.ARIA_ROLES.hasOwnProperty(role) && !axs.constants.ARIA_ROLES[role].abstract);
+        var valid = (axs.constants.ARIA_ROLES.hasOwnProperty(role) && !axs.constants.ARIA_ROLES[role]['abstract']);
         var result = { name: role, valid: valid };
         if (valid) {
             result.details = axs.constants.ARIA_ROLES[role];
@@ -626,3 +626,49 @@ test('composedTreeSearch should ignore distributed nodes if browser does not sup
         ok(true);
     }
 });
+
+module('getImplicitRole', {
+    setup: function() {}
+});
+
+test('get implicit role for button', function() {
+    var element = document.createElement('button');
+    var actual = axs.utils.getImplicitRole(element);
+    equal(actual, 'button');
+});
+
+test('get implicit role for input type=button', function() {
+    var element = document.createElement('input');
+    element.setAttribute('type', 'button');
+    var actual = axs.utils.getImplicitRole(element);
+    equal(actual, 'button');
+});
+
+test('get implicit role for input type=range', function() {
+    var element = document.createElement('input');
+    element.setAttribute('type', 'range');
+    var actual = axs.utils.getImplicitRole(element);
+    equal(actual, 'slider');
+});
+
+test('get implicit role for li out of context', function() {
+    var element = document.createElement('li');
+    var actual = axs.utils.getImplicitRole(element);
+    strictEqual(actual, '');
+});
+
+test('get implicit role for li child of ul', function() {
+    var element = document.createElement('ul');
+    element = element.appendChild(document.createElement('li'));
+    var actual = axs.utils.getImplicitRole(element);
+    equal(actual, 'listitem');
+});
+
+test('get implicit role for li descendant of ul', function() {
+    var element = document.createElement('ul');
+    element = element.appendChild(document.createElement('div'));  // bad html i know but good for test
+    element = element.appendChild(document.createElement('li'));
+    var actual = axs.utils.getImplicitRole(element);
+    strictEqual(actual, '');
+});
+
